fix(netwatch): hold on to netmon sender reference in android (#2923)
## Description

With #2913, the netmon actor properly exits when the channel sender is
dropped, but the android implementation never kept the sender around.
The magicsock actor would then error when trying to subscribe to the
network monitor, and the `Endpoint` would be unusable.

## Change checklist

- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.