chore: prune some deps (#2932)
## Description

Getting started on cleaning up our deps tree as much as possible. 
Closes https://github.com/n0-computer/iroh/issues/2883

This will continue to be an ongoing thing. We have a couple deps that
are pending on releases from further deps. We've managed to clear out
most of our own repos and iroh dependencies.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.

---------

Co-authored-by: dignifiedquire <me@dignifiedquire.com>
Co-authored-by: root <me@n0.computer>