feat!: Add support for TCP (#32)
## Description
Hi. I needed TCP for my application so I added a configuration option to
open a TCP port.

## Breaking Changes
Added a new field to the `Config` struct. 

## Notes & open questions
My router supports NAT-PMP so it would be a good idea for someone to
test UPNP and PCP.

## Change checklist
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.