fix(nap_pmp): reject buffers that are too short on responses (#34)
## Description
Fixes a panic in decoding a response from the nat box due to incorrect
buffer size verifications. We were checking that the buffer size is
within range for all different kinds of responses, but for the specific
map response the buffer size is exact.

Also updates son libraries due to advisories
## Notes & open questions
<!-- Any notes, remarks or open questions you have to make about the PR.
-->
## Change checklist
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.